Fixed bug where Segment used Int for loudnessEnd.  All endpoints that require a URI of a specific type now throw an error if a different type is passed in before any requests are made to the Spotify web API. Added Episodes methods.

+++ Sources/SpotifyWebAPI/Object Model/Audio Analysis Objects/Segment.swift
@@ -70,7 +70,7 @@ public struct Segment: Hashable {
     ///
     /// This value should be equivalent to the `loudnessStart` of the
     /// following segment.
-    public let loudnessEnd: Int?
+    public let loudnessEnd: Double?
 
 }
